Changed: Moved zstd_calc_size_uncached to a common module.
